Move uploads module into status where it is used
